+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Added: `<text>` elements can now have a `rel` attribute pointing to another
+  element which provides the bounding box for the new text element, and as the
+  basis for attributes such as `text-loc`. This feature makes it easier to
+  position multiple text elements relative to an element.
+
 ## [0.27.0 - 2026-02-06]
 
 - Changed: `<line>` based connectors are now axis-aligned by default where